# docs/openapi.yaml
openapi: 3.0.3
info:
  title: 集成与管控子系统 API
  description: |
    集成与管控子系统 REST API 文档

    提供统一身份认证、安全管控、系统集成、运维监控等功能接口。

    **国产化技术栈：**
    - 数据库：达梦 DM8
    - 国密算法：SM2/SM3/SM4
    - 认证：OAuth2.0 + JWT

    **安全特性：**
    - 三权分立（管理员/审计员/操作员）
    - 数据分级管控（L1-L4）
    - 完整审计日志
  version: 1.0.0
  contact:
    name: CTG Integration Team
servers:
  - url: http://localhost:8080/api
    description: 开发环境
...
paths:
...
  /users:
    post:
      tags: [用户管理]
      summary: 创建用户
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '200':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserVO'
    get:
      tags: [用户管理]
      summary: 用户列表
      security:
        - bearerAuth: []
      parameters:
        - name: department
          in: query
          schema:
            type: string
        - name: enabled
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: 用户列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserVO'
  /users/{id}:
    get:
      tags: [用户管理]
      summary: 获取用户
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 用户详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserVO'
    put:
      tags: [用户管理]
      summary: 更新用户
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserVO'
    delete:
      tags: [用户管理]
      summary: 删除用户
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 删除成功
  /users/{id}/status:
    patch:
      tags: [用户管理]
      summary: 切换状态
      description: 启用/禁用用户
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: enabled
          in: query
          required: true
          schema:
            type: boolean
      responses:
        '200':
          description: 状态切换成功
  /users/{id}/roles:
    post:
      tags: [用户管理]
      summary: 分配角色
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                type: integer
                format: int64
      responses:
        '200':
          description: 角色分配成功
...
  /monitor/alerts:
    get:
      tags: [运维监控]
      summary: 告警列表
      security:
        - bearerAuth: []
      parameters:
        - name: severity
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: 告警列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AlertVO'
  /monitor/alerts/{id}/acknowledge:
    post:
      tags: [运维监控]
      summary: 确认告警
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: operator
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 确认成功
  /monitor/statistics:
    get:
      tags: [运维监控]
      summary: 监控统计
      security:
        - bearerAuth: []
      parameters:
        - name: startTime
          in: query
          schema:
            type: string
        - name: endTime
          in: query
          schema:
            type: string
      responses:
        '200':
          description: 统计数据
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MonitorStatisticsVO'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CreateUserRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 64
        password:
          type: string
          minLength: 8
          maxLength: 32
        realName:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        employeeId:
          type: string
        department:
          type: string
        roleIds:
          type: array
          items:
            type: integer
            format: int64
    UpdateUserRequest:
      type: object
      properties:
        realName:
          type: string
        email:
          type: string
        phone:
          type: string
        department:
          type: string
        roleIds:
          type: array
          items:
            type: integer
            format: int64
    UserVO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        realName:
          type: string
        email:
          type: string
        phone:
          type: string
        employeeId:
          type: string
        department:
          type: string
        enabled:
          type: boolean
        lastLoginAt:
          type: string
          format: date-time
        roles:
          type: array
          items:
            $ref: '#/components/schemas/RoleVO'
        createdAt:
          type: string
          format: date-time
    RoleVO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        code:
          type: string
        name:
          type: string
        description:
          type: string
        roleType:
          type: string
          enum: [ADMIN, AUDITOR, OPERATOR]
        permissions:
          type: array
          items:
            $ref: '#/components/schemas/PermissionVO'
    PermissionVO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        code:
          type: string
        name:
          type: string
        type:
          type: string
          enum: [MENU, BUTTON, API, DATA]
        resource:
          type: string
        method:
          type: string
...
    AlertVO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        alertName:
          type: string
        alertType:
          type: string
        severity:
          type: string
          enum: [INFO, WARNING, CRITICAL, FATAL]
        status:
          type: string
          enum: [TRIGGERED, ACKNOWLEDGED, RESOLVED]
        message:
          type: string
        source:
          type: string
        metricName:
          type: string
        metricValue:
          type: number
          format: double
        threshold:
          type: number
          format: double
        triggeredAt:
          type: string
          format: date-time
        acknowledgedAt:
          type: string
          format: date-time
        acknowledgedBy:
          type: string
        resolvedAt:
          type: string
          format: date-time
    MonitorStatisticsVO:
      type: object
      properties:
        totalAlerts:
          type: integer
          format: int64
        activeAlerts:
          type: integer
          format: int64
        resolvedAlerts:
          type: integer
          format: int64
        avgResponseTime:
          type: number
          format: double
        systemUptime:
          type: number
          format: double
        alertCountsByType:
          type: array
          items:
            type: object
            properties:
              type:
                type: string
              count:
                type: integer
                format: int64
        alertCountsBySeverity:
          type: array
          items:
            type: object
            properties:
              severity:
                type: string
              count:
                type: integer
                format: int64