# src/main/java/com/ctg/integration/dto/user/UpdateUserRequest.java
package com.ctg.integration.dto.user;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.util.List;

/**
 * 更新用户请求DTO
 *
 * @author CTG
 * @since 2026-07-01
 */
@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class UpdateUserRequest {

    private String realName;
    private String email;
    private String phone;
    private String department;
    private List<Long> roleIds;
}